Remove single-implementation interfaces for services
ScoringServiceImpl => ScoringService
AuthorizationServiceImpl => AuthorizationService

Unused imports also removed during refactoring.
