+++ entity-management-web/src/main/java/eu/europeana/entitymanagement/web/EMController.java
@@ -16,9 +16,8 @@ import eu.europeana.entitymanagement.vocabulary.EntityProfile;
 import eu.europeana.entitymanagement.vocabulary.FormatTypes;
 import eu.europeana.entitymanagement.vocabulary.WebEntityConstants;
 import eu.europeana.entitymanagement.web.model.EntityPreview;
-import eu.europeana.entitymanagement.web.service.impl.EntityRecordService;
-import eu.europeana.entitymanagement.web.service.impl.EntityRecordUtils;
-import eu.europeana.entitymanagement.web.service.impl.MetisDereferenceService;
+import eu.europeana.entitymanagement.web.service.EntityRecordService;
+import eu.europeana.entitymanagement.web.service.MetisDereferenceService;
 import io.swagger.annotations.ApiOperation;
 import java.util.Date;
 import java.util.Optional;
