+++ entity-management-web/src/main/java/eu/europeana/entitymanagement/web/service/EntityObjectFactory.java
@@ -1,4 +1,4 @@
-package eu.europeana.entitymanagement.web.service.impl;
+package eu.europeana.entitymanagement.web.service;
 
 import eu.europeana.entitymanagement.definitions.model.Entity;
 import eu.europeana.entitymanagement.definitions.model.impl.AgentImpl;
